Support storing user credentials as bcrypt hashes
Instead of storing user credentials as plaintext passwords, support
bcrypt hashes.  This is backwards compatible in that it will only use
the hash if it's available, but will fall back to plaintext if not.
Also, during startup a log entry will be generated if there are any
users that have plaintext passwords configured.